fix: expand player state to 64-bit integer (#1361)
* Expand player state to 64-bit integer

Changed player state variables and related methods from uint32_t to uint64_t to support additional state flags for newer client versions. Also updated protocol parsing and getter/setter signatures accordingly. Minor fix in UIWidget HTML layout to use std::max<int> for clarity.

* Refactor player state handling to support 64-bit masks

Updated player state logic to use 64-bit masks for better extensibility. Replaced bitwise operations with new helper functions for checking and iterating state changes. Updated related Lua and C++ code to use the new approach, improving maintainability and future-proofing state management.

+++ src/client/const.h
@@ -268,7 +268,7 @@ namespace Otc
         NpcIconTradeQuest
     };
 
-    enum PlayerStates : uint32_t
+    enum PlayerStates : uint64_t
     {
         IconNone = 0,
         IconPoison = 1,
